[Fix] 주문 생성 api 수정 - #161

+++ src/main/java/com/nhnacademy/heukbaekbookshop/book/dto/response/book/BookSummaryResponse.java
@@ -7,6 +7,6 @@ public record BookSummaryResponse(Long id,
                                   boolean isPackable,
                                   BigDecimal price,
                                   BigDecimal salePrice,
-                                  double discountRate,
+                                  BigDecimal discountRate,
                                   String thumbnailUrl) {
 }
