+++ src/main/java/com/nhnacademy/heukbaekbookshop/book/dto/response/book/BookViewResponse.java
@@ -4,6 +4,7 @@ import com.nhnacademy.heukbaekbookshop.category.dto.response.CategorySummaryResp
 import com.nhnacademy.heukbaekbookshop.contributor.dto.response.ContributorSummaryResponse;
 import com.nhnacademy.heukbaekbookshop.contributor.dto.response.PublisherSummaryResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record BookViewResponse(Long id,
@@ -15,7 +16,7 @@ public record BookViewResponse(Long id,
                                boolean isPackable,
                                String price,
                                String salePrice,
-                               float discountRate,
+                               BigDecimal discountRate,
                                long popularity,
                                String status,
                                String thumbnailUrl,
